--- src/pos.html
@@ -1,56 +1,0 @@
-<!--
-  ~ Copyright (c) 2019. Author Hubert Formin <hformin@gmail.com>
-  -->
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Print preview</title>
-    <style>
-        @page {
-            margin: 0;
-            page-break-before: always;
-        }
-        body {
-            font-size: 12px;
-            /*line-height: 24px;*/
-        }
-        .barcode-container::after{content: '';display: table;clear: both;}
-        .font {font-size: 12px; margin: 0 15px;}
-
-        table {
-            width: 100%;
-            display: table;
-            border-collapse: collapse;
-            border-spacing: 0;
-            font-family: monospace, cursive;
-        }
-
-        table tbody {
-            border-top: 1px solid #999;
-        }
-
-        table th, table td {
-            padding: 7px 2px;
-        }
-
-        table tr {
-            border-bottom: 1px dotted #999;
-            padding: 5px 0;
-            text-align: center;
-        }
-
-        table img {
-            max-height: 40px;
-        }
-    </style>
-</head>
-<body>
-<section id="main"></section>
-<!--<script src="libs/jquery.min.js"></script>-->
-<!--<script src="libs/jsbarcode.min.js"></script>-->
-<!--<script src="libs/qrcode.min.js"></script>-->
-<script>let exports = {};</script>
-<script type="module" src="lib/body-init.js"></script>
-</body>
-</html>
