+++ src/renderer/index.html
@@ -47,10 +47,5 @@
 </head>
 <body>
 <section id="main"></section>
-<!--<script src="libs/jquery.min.js"></script>-->
-<!--<script src="libs/jsbarcode.min.js"></script>-->
-<!--<script src="libs/qrcode.min.js"></script>-->
-<script>let exports = {};</script>
-<script type="module" src="lib/body-init.js"></script>
 </body>
 </html>
